requirements.txt
toml>=0.10.2
requests==2.32.3
numpy==1.26.4
python-dotenv==1.0.1
sentence-transformers==3.4.1
faiss-cpu==1.9.0.post1
pydantic==2.10.6
httpx==0.27.2
gradio==5.23.1
langgraph==0.2.62
langchain_groq==0.2.4
langchain_core==0.3.80
rank_bm25==0.2.2
langgraph-cli==0.1.79
duckduckgo-search>=6.0.0
boto3>=1.34.0
langchain-aws>=0.1.0